requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
coverage==7.3.2
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
flake8==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
smsaero-api==2.2.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
gunicorn
